requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai==1.3.0
cohere==5.5.4
qdrant-client==1.9.1
python-dotenv==1.0.0
pytest==8.3.2
pytest-asyncio==0.21.1
httpx==0.25.2